+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- The `hotg-rune-wasmer-runtime` and `hotg-rune-wasm3-runtime` crates have been
+  merged into `hotg-rune-runtime`
 ## [0.10.0] - 2021-10-25
 
 ### Added